+++ apps/react-test-app/src/routes/~__root.tsx
@@ -6,7 +6,7 @@ export const Route = createRootRoute({
 
 function RootComponent() {
   return (
-    <div className="flex h-screen w-full flex-col items-center justify-center gap-2 overflow-y-auto bg-slate-50 text-slate-950 subpixel-antialiased">
+    <div className="flex h-screen w-full flex-col items-center justify-center gap-2 overflow-y-auto bg-slate-50 text-slate-950 subpixel-antialiased dark:bg-slate-950 dark:text-slate-50">
       <Outlet />
     </div>
   )